Fixed a floating-point bug
We ended up with negative values being computed, since the database stored
single precision floating point, where as JavaScript computes exclusively using
double precision (i.e. 32-bit) floats.

In other words, we get a value of 1.99, but the database would store the
value as 2, since it would round the value.

If the newly retrieved inputed value never changed from the last, then we get
an arithmetic of

    kwh - old_kwh < 0

Where kwh = 1.99 represents the data retrieved, and old_kwh = 2 represents the
data in the database.

+++ migrations/001-devices.js
@@ -38,8 +38,8 @@ exports.up = function(next) {
   'CREATE TABLE IF NOT EXISTS `hourly_totals` ( \
     `id` int(11) unsigned NOT NULL AUTO_INCREMENT, \
     `time` datetime NOT NULL, \
-    `start_kwh` float NOT NULL, \
-    `hour_kwh` float NOT NULL, \
+    `start_kwh` double NOT NULL, \
+    `hour_kwh` double NOT NULL, \
     `device_id` int(11) unsigned NOT NULL, \
     PRIMARY KEY (`id`), \
     KEY `device_id` (`device_id`) \
